Явное приведение типов в make_interval
Postgres не выводит тип параметра в именованном аргументе, из-за чего
падала постановка уведомления в очередь и фоновое обслуживание.

+++ api/collect.js
@@ -167,7 +167,7 @@ async function calculatorState(sessionId) {
 async function enqueue(client, { sessionId, kind, dedupeKey, text, delaySec = 0 }) {
   await client.query(
     `INSERT INTO notifications (session_id, kind, dedupe_key, text, send_after)
-     VALUES ($1, $2, $3, $4, now() + make_interval(secs => $5))
+     VALUES ($1, $2, $3, $4, now() + make_interval(secs => $5::int))
      ON CONFLICT (dedupe_key) DO NOTHING`,
     [sessionId, kind, dedupeKey || null, text, delaySec],
   );
